v0.3.0 General style guide
 RWD navigation,
 additional configs at theme,
 Separate file NavigationBlock and PageWrapperBlock
